fix build issues for docs
