+++ packages/tools/src/tools/annotation/LengthTool.ts
@@ -38,7 +38,6 @@ import {
 } from '../../cursors/elementCursor'
 
 import {
-  Annotation,
   EventTypes,
   ToolHandle,
   TextBoxHandle,
@@ -46,31 +45,7 @@ import {
   ToolProps,
   InteractionTypes,
 } from '../../types'
-
-interface LengthAnnotation extends Annotation {
-  data: {
-    handles: {
-      points: Types.Point3[]
-      activeHandleIndex: number | null
-      textBox: {
-        hasMoved: boolean
-        worldPosition: Types.Point3
-        worldBoundingBox: {
-          topLeft: Types.Point3
-          topRight: Types.Point3
-          bottomLeft: Types.Point3
-          bottomRight: Types.Point3
-        }
-      }
-    }
-    label: string
-    cachedStats: {
-      [targetId: string]: {
-        length: number
-      }
-    }
-  }
-}
+import { LengthAnnotation } from '../../types/ToolSpecificAnnotationTypes'
 
 /**
  * LengthTool let you draw annotations that measures the length of two drawing
